pyproject.toml
[project]
name = "multilingual-layer-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.0"
dependencies = [
    "cuml-cu12>=25.12.0",
    "cupy-cuda12x>=13.6.0",
    "datasets>=4.5.0",
    "dotenv>=0.9.9",
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "nltk>=3.9.2",
    "omegaconf>=2.3.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.10.0",
    "tqdm>=4.67.1",
    "transformers>=5.0.0",
    "vllm>=0.2.5",
]
